requirements.txt
flask==3.1.0
gunicorn==23.0.0
google-cloud-bigquery==3.27.0
pandas==2.2.3
pandas-gbq==0.24.0
pyarrow==18.1.0
db-dtypes==1.3.1
